requirements.txt
ollama
duckdb
streamlit
watchdog
flask
markdown-it-py

isort
black
